requirements.txt
google-api-core==1.21.0
google-api-python-client==1.9.3
google-auth==1.19.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
google-cloud-bigquery==1.25.0
google-cloud-core==1.3.0
google-resumable-media==0.5.1
googleapis-common-protos==1.52.0
httplib2==0.18.1
matplotlib==3.2.2
oauth2client==4.1.3
oauthlib==3.1.0
openpyxl==3.0.4